different refactorings

+++ MoneyManager/MoneyManager.Shared/DataAccess/AbstractDataAccess.cs
@@ -1,7 +1,12 @@
 ﻿using BugSense;
+using MoneyManager.OperationContracts;
+using MoneyManager.Src;
+using MoneyManager.ViewModels;
 using System;
+using System.Threading.Tasks;
+using Windows.UI.Popups;
 
-namespace MoneyManager.ViewModels
+namespace MoneyManager.DataAccess
 {
     public abstract class AbstractDataAccess<T> : IDataAccess<T>
     {
@@ -25,16 +30,32 @@ namespace MoneyManager.ViewModels
             }
         }
 
-        public void Delete(T itemToDelete)
+        public async void Delete(T itemToDelete)
         {
             try
             {
-                DeleteFromDatabase(itemToDelete);
+                if (await IsDeletionConfirmed())
+                {
+                    DeleteFromDatabase(itemToDelete);
+                }
             }
             catch (Exception ex)
             {
                 BugSenseHandler.Instance.LogException(ex);
             }
+        }
+
+        private async Task<bool> IsDeletionConfirmed()
+        {
+            var dialog = new MessageDialog(Utilities.GetTranslation("DeleteEntryQuestionMessage"),
+                Utilities.GetTranslation("DeleteQuestionMessage"));
+            dialog.Commands.Add(new UICommand(Utilities.GetTranslation("YesLabel")));
+            dialog.Commands.Add(new UICommand(Utilities.GetTranslation("NoLabel")));
+            dialog.DefaultCommandIndex = 1;
+
+            var result = await dialog.ShowAsync();
+
+            return result.Label == Utilities.GetTranslation("YesLabel");
         }
 
         public void LoadList()
